Patches a duplication bug in single dimen
Additionally it also adds tabbing in the dimen template for for prettyfying.
