+++ resources/META-INF/plugin.xml
@@ -1,7 +1,7 @@
 <idea-plugin version="2">
     <id>com.hr.dimenify</id>
     <name>Dimenify</name>
-    <version>3.2</version>
+    <version>3.3</version>
     <vendor email="anvithv4@gmail.com" url="https://www.anvith.in">Anvith Bhat</vendor>
 
     <description><![CDATA[
@@ -13,6 +13,9 @@
     ]]></description>
 
     <change-notes><![CDATA[
+    <b>v3.3</b><br/>
+        - Fixed a long standing bug in single dimen conversion.
+        - Prettier xml
      <b>v3.2</b><br/>
         - Fixed a message formatter bug causing floated strings to be prettified with comma.
