+++ src/com/hr/dimenify/action/GenerateSingleDimenAction.java
@@ -8,30 +8,30 @@ import com.intellij.openapi.actionSystem.LangDataKeys;
 import com.intellij.openapi.actionSystem.PlatformDataKeys;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.fileTypes.StdFileTypes;
-import com.intellij.psi.PsiDirectory;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiFile;
 import com.intellij.psi.impl.source.xml.XmlAttributeImpl;
 import com.intellij.psi.impl.source.xml.XmlDocumentImpl;
 import com.intellij.psi.impl.source.xml.XmlTagImpl;
 import com.intellij.psi.impl.source.xml.XmlTextImpl;
+import com.intellij.psi.xml.XmlFile;
+import com.intellij.psi.xml.XmlTag;
 
 import javax.swing.*;
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.text.MessageFormat;
 import java.util.ArrayList;
-import java.util.Locale;
+import java.util.HashMap;
 
 import static com.hr.dimenify.util.Constants.ERROR_CODE;
 
 
 public class GenerateSingleDimenAction extends AbstractDimenAction {
 
+    String attributeName;
+
+
     @Override
     public void actionPerformed(AnActionEvent e) {
         super.actionPerformed(e);
-        PsiFile psiFile = e.getData(LangDataKeys.PSI_FILE);
+        psiFile = e.getData(LangDataKeys.PSI_FILE);
         Editor editor = PlatformDataKeys.EDITOR.getData(e.getDataContext());
 
         if (psiFile == null || editor == null) {
@@ -42,6 +42,7 @@ public class GenerateSingleDimenAction extends AbstractDimenAction {
         int offset = editor.getCaretModel().getOffset();
         PsiElement psiElement = psiFile.findElementAt(offset);
         if (psiFile.getFileType() == StdFileTypes.XML) {
+            xmlFile = (XmlFile) psiFile;
             currentBucketIndex = getBucketIndex(psiFile);
             if (currentBucketIndex == -1) {
                 if (data.size() == 10) {
@@ -70,37 +71,40 @@ public class GenerateSingleDimenAction extends AbstractDimenAction {
                 }
 
             }
-            values = getInsertionValuesElement(psiElement);
-            if (values == null) {
-                e.getPresentation().setEnabled(false);
-                return;
-            } else {
-                if (psiFile.getParent().getParent() != null) {
-                    PsiDirectory psiDirectory = psiFile.getParent().getParent();
-                    createDirectoriesAndFilesIfNeeded(psiDirectory, Constants.Mode.SINGLE);
-
-                }
-            }
+            initialize(psiElement);
         } else {
             e.getPresentation().setEnabled(false);
             return;
         }
     }
 
-    @Override
-    protected void writeBulkValuesToFiles() {
+    protected void calculateAndWriteScaledValueToFiles() {
+        XmlTag[] dimens = getDimenValuesInFile(xmlFile);
+        for (XmlTag tag : dimens) {
+            try {
+                if (tag.getAttribute("name").getValue().equalsIgnoreCase(attributeName)) {
+                    dimens = new XmlTag[]{tag};
+                    break;
+                }
+            } catch (Exception e) {
+                return;
+            }
+        }
 
+        if (dimens != null && dimens.length > 0) {
+            HashMap<String, Float>[] floatDimen = normalizeToHashMap(dimens, currentBucketIndex);
+            writeScaledValuesToFiles(psiFile.getParent().getParent(), floatDimen);
+        }
     }
 
 
-    private String[] getInsertionValuesElement(PsiElement psiElement) {
+    private void initialize(PsiElement psiElement) {
         PsiElement selectedNode = psiElement;
         PsiElement rootParent;
         PsiElement subNode;
         if (psiElement != null && psiElement.getParent() != null && psiElement.getParent().getParent() != null) {
             subNode = psiElement.getParent();
             rootParent = subNode.getParent();
-            ArrayList<PsiElement> elementHierachy = new ArrayList<>();
 
             while (!(rootParent instanceof XmlDocumentImpl) && (!(selectedNode instanceof XmlTagImpl) || !((XmlTagImpl) selectedNode).getName().equals(Constants.DIMEN_TAG))) {
                 rootParent = rootParent.getParent();
@@ -122,61 +126,32 @@ public class GenerateSingleDimenAction extends AbstractDimenAction {
                     }
 
                     if (attribute != null) {
-                        String attributeName = attribute.getValue();
+                        attributeName = attribute.getValue();
                         String val = value.getValue().toLowerCase().trim();
-
                         if (val.endsWith(Constants.DP) || val.endsWith(Constants.SP)) {
-
-                            return showScaleDialog(attributeName, val, val.endsWith(Constants.DP));
+                            showScaleDialog(val.endsWith(Constants.DP));
                         }
 
                     }
                 }
             }
-            return null;
 
-        } else {
-            return null;
         }
     }
 
-    private String[] showScaleDialog(String attributeName, String val, boolean isDp) {
+    private void showScaleDialog(boolean isDp) {
         SingleDimenDialog singleDimenDialog = new SingleDimenDialog(project, isDp, data);
         singleDimenDialog.show();
         int invalidIndex = singleDimenDialog.invalidBucketIndex();
         if (singleDimenDialog.isOK() && (invalidIndex == -1 || (invalidIndex == ERROR_CODE[1] && data.size() == Constants.MAX_DIMENS))) {
-            DecimalFormatSymbols otherSymbols = new DecimalFormatSymbols(Locale.getDefault());
-            otherSymbols.setDecimalSeparator('.');
-            otherSymbols.setGroupingSeparator(',');
-            DecimalFormat formatter = new DecimalFormat("#0.0", otherSymbols);
+
             ArrayList<Dimen> data = singleDimenDialog.getConversionValues();
             saveValues(data);
+            createDirectoriesAndFilesIfNeeded(psiFile.getParent().getParent());
 
-
-            float scaleFactor[] = new float[data.size()];
-            for (int i = 0; i < data.size(); i++) {
-                scaleFactor[i] = isDp ? data.get(i).getFactorDp() : data.get(i).getFactorSp();
-            }
-            float mdpiValue = 0;
-            try {
-                mdpiValue = Float.parseFloat(val.substring(0, val.length() - 2)) / scaleFactor[currentBucketIndex];
-            } catch (NullPointerException | NumberFormatException e) {
-                return null;
-            }
-            float scaledValues[] = new float[scaleFactor.length];
-            String elementsScaled[] = new String[scaleFactor.length];
-
-            for (int i = 0; i < scaledValues.length; i++) {
-                scaledValues[i] = mdpiValue * scaleFactor[i];
-                elementsScaled[i] = MessageFormat.format(Constants.PLACEHOLDER_DIMEN, attributeName
-                        , formatter.format(scaledValues[i])
-                        , val.endsWith(Constants.DP) ? Constants.DP : Constants.SP);
-            }
-            return elementsScaled;
         } else if (singleDimenDialog.isOK()) {
             singleDimenDialog.showAlert(invalidIndex);
         }
-        return null;
     }
